refactor: 심사 거절 시, 즉시 삭제(#417)
